Dockerfile: Use unversioned golang:alpine image
The Dockerfile hardcodes the use of golang:1.16-alpine.

Both golang 1.16 and the alpine version used in this image are
unmaintained at this point, with most likely some known unfixed security
issues.

This PR switches to an unversioned golang/alpine image, which should
ensure we always use an up-to-date image.
We can switch back to a versioned image if this causes issues and it
turns out we need more predictability on the golang version used.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>